Add error handling to GetLogLevelByName method
